+++ .github/workflows/ruff.yaml
@@ -29,6 +29,6 @@ jobs:
         cd stgraph/benchmark_tools
         ruff check .
         cd ../../
-        cd stgraph/nn
-        ruff check .
-        cd ../../
+        cd stgraph/nn/pytorch
+        ruff check gcn_conv.py
+        cd ../../../
