+++ examples/README.md
@@ -12,7 +12,7 @@ Open up your favourite text editor or Python IDE and create a file named `model.
 
 ```python
 import torch.nn as nn
-from stgraph.nn.pytorch.gcn_conv import GCNConv
+from stgraph.nn.pytorch.static.gcn_conv import GCNConv
 
 
 class GCN(nn.Module):
