+++ stgraph/nn/pytorch/static/__init__.py
@@ -1,0 +1,1 @@
+"""State-of-the-art Static Graph Neural Networks written for PyTorch backend."""
